Back/requirements.txt
flask
flask-cors
pandas
scikit-learn
psycopg2-binary